Make the bull queue library a soft dependency
For applications which don't use the `bull`-based persistent event queueing system, rudderstack-sdk-node still installs and requires the module. `bull` is kind of heavyweight -- it includes a bunch of redis dependencies and a `msgpackr` library which has a native extension. For my app this native extension has been rather annoying to keep support for despite us not using it at all.

I think it'd be a better default to not require the `bull` library until the sdk actually needs it to improve startup time, and if you folks are open to it, not listing it as a dependency in `package.json` as it is really an optional dependency. The docs would have to be updated to note that users who wanted to use the persistent queue would need to install `bull` separately. I don't think that's too wild.

Also, fwiw, bull has been deprecated in favour of `bullmq` which is a bit better supported these days, which is maybe another reason not to pull it into all users' apps.